+++ src/roi/roiSignal.m
@@ -1,16 +1,20 @@
-function [signal, xpts] = roiSignal(imStack, roiMask, sampleRate)
+function [signal, xpts] = roiSignal(imStack, roiMask, sampleRate, bkgdWindow, medianFlag)
     % ROISIGNAL
     %
     % Syntax:
     %   roiSignal(imStack, roiMask, frameRate);
     % 
     % Inputs:
-    %   imStack     3D matrix - [X, Y, T]
+    %   imStack         3D matrix - [X, Y, T]
     %       Raw imaging data stack
-    %   roiMask     binary 2D matrix [x, Y]
+    %   roiMask         binary 2D matrix [x, Y]
     %       Mask of designating ROI 
     %   sampleRate      numeric (default = 25)
     %       Samples/frames per second (Hz)
+    %   bkgdWindow      vector [1 x 2]
+    %       Start and stop frames for background estimate, returns dF/F
+    %   medianFlag      logical (default = false)
+    %       Use median flag for background estimation instead of mean
     %
     % Outputs:
     %   signal      vector - [1, T]
@@ -19,18 +23,30 @@ function [signal, xpts] = roiSignal(imStack, roiMask, sampleRate)
     %       Time points associated with signal 
     %
     % See also:
-    %   ROISIGNALPLOT
+    %   ROISIGNALPLOT, ROISIGNALS
     % 
     % History:
     %   22Aug2020 - SSP
+    %   02Dec2020 - SSP - Added bkgd estimation options
     % --------------------------------------------------------------------
-
-    if nargin < 3
-        sampleRate = 25;  % Hz
-    end
-
+    
     xpts = 1/sampleRate : 1/sampleRate : size(imStack, 3)/sampleRate;
 
     [a, b] = find(roiMask == 1);
-    signal = imStack(a, b, :);
-    signal = squeeze(mean(mean(signal, 1), 2));
+    signal = zeros(numel(a), size(imStack, 3));
+    for i = 1:numel(a)
+        signal(i, :) = squeeze(imStack(a(i), b(i), :));
+    end
+    signal = mean(signal);
+    
+    if nargin >= 4 && ~isempty(bkgdWindow)
+        if nargin < 5
+            medianFlag = false;
+        end
+        if medianFlag
+            bkgd = median(signal(bkgdWindow(1):bkgdWindow(2)));
+        else
+            bkgd = mean(signal(bkgdWindow(1):bkgdWindow(2)));
+        end
+        signal = (signal - bkgd) / bkgd;
+    end